Use true lambertian reflection
Selecting a random point within the unit sphere.
Indeed there is a high probability to pick directions close to the
normal. The distribution scales by cos³(teta) with teta the angle from
the normal.

A lambertian distribution has a distribution of cos(teta).
True Lambertian has the probability higher for ray scattering close to
the normal, but the distribution is more uniform.
This is done by picking points on the surface of the unit sphere.